M2T2/requirements.txt
h5py
hydra-core
matplotlib
meshcat
numpy
scikit-learn
scipy
tensorboard
torch
torchvision
trimesh